refactor: Delete unnecessary files, Organize files, and standardized folder names to start with lowercase
- 기능 수정 없음
