+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:theme="@style/Theme.Map"
         tools:targetApi="31">
         <activity
-            android:name=".HomeMapActivity"
+            android:name=".activity.HomeMapActivity"
             android:allowBackup="true"
             android:exported="true">
             <intent-filter>
@@ -25,7 +25,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".DataSearchActivity"
+            android:name=".activity.DataSearchActivity"
             android:exported="true"
             android:windowSoftInputMode="adjustResize">
         </activity>
