+++ app/src/main/java/campus/tech/kakao/map/activity/HomeMapActivity.kt
@@ -1,25 +1,22 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.activity
 
 import android.annotation.SuppressLint
 import android.content.Intent
-import android.content.pm.PackageManager
 import android.os.Bundle
-import android.util.Base64
-import android.util.Log
 import android.widget.EditText
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
+import campus.tech.kakao.map.R
 import com.kakao.vectormap.KakaoMap
 import com.kakao.vectormap.KakaoMapReadyCallback
 import com.kakao.vectormap.MapView
 import com.kakao.vectormap.KakaoMapSdk
 import com.kakao.vectormap.MapLifeCycleCallback
 import java.lang.Exception
-import java.security.MessageDigest
 
 class HomeMapActivity : AppCompatActivity() {
-    lateinit var mapView: MapView
-    lateinit var searchBar: EditText
+    private lateinit var mapView: MapView
+    private lateinit var searchBar: EditText
 
     @SuppressLint("MissingInflatedId")
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -31,14 +28,15 @@ class HomeMapActivity : AppCompatActivity() {
         KakaoMapSdk.init(this, key)
 
         mapView = findViewById(R.id.mapView)
-        mapView.start(object: MapLifeCycleCallback(){
+        //KaKao Map UI에 띄우기
+        mapView.start(object : MapLifeCycleCallback() {
             override fun onMapDestroy() {
             }
 
             override fun onMapError(p0: Exception?) {
             }
 
-        }, object: KakaoMapReadyCallback(){
+        }, object : KakaoMapReadyCallback() {
             override fun onMapReady(p0: KakaoMap) {
             }
