+++ app/src/main/java/campus/tech/kakao/map/data/RecentSearchData.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map.Data
+package campus.tech.kakao.map.data
 
 
 data class RecentSearchData(val name: String, val address: String, val time: Long)
