+++ app/src/main/java/campus/tech/kakao/map/data/SearchData.kt
@@ -1,3 +1,4 @@
-package campus.tech.kakao.map.Data
+package campus.tech.kakao.map.data
 
+//2주차 PR 피드백을 위해 남겨 놓음
 data class SearchData(val name: String, val category: String, val address: String)
