+++ app/src/main/java/campus/tech/kakao/map/dataRepository/SearchDataRepository.kt
@@ -1,9 +1,7 @@
-package campus.tech.kakao.map.DataRepository
+package campus.tech.kakao.map.dataRepository
 
-import android.content.Context
 import android.util.Log
 import campus.tech.kakao.map.BuildConfig
-import campus.tech.kakao.map.Data.SearchData
 import campus.tech.kakao.map.retrofit.MapSearchResponse
 import campus.tech.kakao.map.retrofit.RetrofitService
 import retrofit2.Call
@@ -12,9 +10,7 @@ import retrofit2.Response
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 
-class SearchDataRepository(context: Context) {
-    val searchResultList: MutableList<SearchData> = mutableListOf()
-
+class SearchDataRepository{
     //실제 사용될 때 Retrofit 객체 생성
     private val retrofitService: RetrofitService by lazy {
         Log.d("yeong", "Use Retrofit!")
