+++ app/src/main/res/layout/activity_data_search.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".DataSearchActivity">
+    tools:context=".activity.DataSearchActivity">
 
     <EditText
         android:id="@+id/searchBar"
