+++ app/src/main/res/layout/activity_home_map.xml
@@ -5,7 +5,7 @@
     android:id="@+id/homeMap"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".HomeMapActivity">
+    tools:context=".activity.HomeMapActivity">
 
     <com.kakao.vectormap.MapView
         android:id="@+id/mapView"
